requirements.txt
flask
flask-cors
pillow
numpy<2.0
tflite-runtime
requests